pyproject.toml
[tool.poetry]
name = "chess-game-stockfish"
version = "0.1.0"
description = ""
authors = ["S1avv"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
clarifai-grpc = "^10.6.6"
flask-cors = "^4.0.1"
python-tgpt = "^0.7.4"
g4f = "^0.3.2.2"
python-chess = "^1.999"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
